refactor: move pull-requests into src/features/pull-requests
- Moves components (PRCard, PRList, Filters), queries (useGitHubPRs,
  useCheckContexts, useViewer) and stores (prStore) into
  src/features/pull-requests/{components,queries,stores}/
- Converts all default exports to named exports for consistency with
  the branches feature slice
- Adds exports.ts barrel (Filters, PRList)
- Updates all import paths to @/ alias
