+++ src/features/branches/components/BranchList/BranchList.tsx
@@ -1,9 +1,9 @@
 import { useQueries } from '@tanstack/react-query'
 import { Copy, ExternalLink, FolderPlus, MoreVertical, RefreshCw, Trash2, X } from 'lucide-react'
 import { useMemo, useState } from 'react'
-import { usePRStore } from '@/store/prStore'
+import { usePRStore } from '@/features/pull-requests/stores/prStore'
 import { useBranchStore } from '../../stores/branchStore'
-import { usePullRequests } from '@/hooks/useGitHubPRs'
+import { usePullRequests } from '@/features/pull-requests/queries/useGitHubPRs'
 import { CopyWithFeedback } from '@/components/shared/CopyWithFeedback'
 import type { Worktree } from '@/types/worktree'
 import type { PullRequest } from '@/types/github'
