+++ src/features/pull-requests/components/PRCard/PRCard.test.tsx
@@ -1,11 +1,11 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import PRCard from './PRCard'
-import { usePRStore } from '../../store/prStore'
-import type { PullRequest, ReviewState } from '../../types/github'
+import { PRCard } from './PRCard'
+import { usePRStore } from '@/features/pull-requests/stores/prStore'
+import type { PullRequest, ReviewState } from '@/types/github'
 
-vi.mock('../../hooks/useCheckContexts', () => ({
+vi.mock('@/features/pull-requests/queries/useCheckContexts', () => ({
   useCheckContexts: vi.fn(() => ({ checks: [], isLoading: false })),
 }))
