+++ src/features/pull-requests/exports.ts
@@ -1,0 +1,2 @@
+export { Filters } from './components/Filters/Filters'
+export { PRList } from './components/PRList/PRList'
