+++ src/features/pull-requests/queries/useCheckContexts.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
-import { createClient, PR_CHECK_CONTEXTS_QUERY } from '../lib/github'
-import { useAuthStore } from '../store/authStore'
-import type { CheckRunContext, StatusContextItem } from '../types/github'
+import { createClient, PR_CHECK_CONTEXTS_QUERY } from '@/lib/github'
+import { useAuthStore } from '@/store/authStore'
+import type { CheckRunContext, StatusContextItem } from '@/types/github'
 
 interface CheckContextsResult {
   node: {
