+++ src/lib/prFilters.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest'
 import { applyFilters } from './prFilters'
 import type { PullRequest } from '../types/github'
-import type { PRFilters } from '../store/prStore'
+import type { PRFilters } from '../features/pull-requests/stores/prStore'
 
 const makePR = (login: string): PullRequest =>
   ({
