+++ src/pages/DashboardPage.tsx
@@ -1,11 +1,10 @@
 import { useState, useEffect } from 'react'
 import { LogOut, Moon, Search, Sun, X } from 'lucide-react'
 import { useAuthStore } from '../store/authStore'
-import { usePRStore } from '../store/prStore'
+import { usePRStore } from '@/features/pull-requests/stores/prStore'
 import { useTheme } from '../hooks/useTheme'
 import { useFeatures } from '../lib/features'
-import Filters from '../components/Filters/Filters'
-import PRList from '../components/PRList/PRList'
+import { Filters, PRList } from '@/features/pull-requests/exports'
 import { BranchList } from '@/features/branches/exports'
 import Footer from '../components/Footer/Footer'
 import { useUpdateCheck, isNewer } from '../hooks/useUpdateCheck'
